dom/indexedDB/test/test_filehandle_truncate.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<html>
<head>
  <title>File Handle Test</title>

  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>

  <script type="text/javascript;version=1.7">
  function* testSteps()
  {
    const name = window.location.pathname;

    var testBuffer = getRandomBuffer(100000);

    let request = indexedDB.open(name, 1);
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    let event = yield undefined;

    let db = event.target.result;
    db.onerror = errorHandler;

    request = db.createMutableFile("test.bin");
    request.onerror = errorHandler;
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    let mutableFile = event.target.result;
    mutableFile.onerror = errorHandler;

    let fileHandle = mutableFile.open("readwrite");
    request = fileHandle.write(testBuffer);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    is(fileHandle.location, 100000, "Correct location");

    for (let i = 0; i < 10; i++) {
      let location = fileHandle.location - 10000;
      fileHandle.location = location;

      request = fileHandle.truncate();
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      is(fileHandle.location, location, "Correct location");

      request = fileHandle.getMetadata({ size: true });
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      is(event.target.result.size, location, "Correct size");
    }

    request = fileHandle.write(testBuffer);
    request.onsuccess = grabEventAndContinueHandler;
    event = yield undefined;

    let location = fileHandle.location;
    for (let i = 0; i < 10; i++) {
      location -= 10000;

      request = fileHandle.truncate(location);
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      is(fileHandle.location, location, "Correct location");

      request = fileHandle.getMetadata({ size: true });
      request.onsuccess = grabEventAndContinueHandler;
      event = yield undefined;

      is(event.target.result.size, location, "Correct size");
    }

    finishTest();
  }
  </script>
  <script type="text/javascript;version=1.7" src="file.js"></script>
  <script type="text/javascript;version=1.7" src="helpers.js"></script>

</head>

<body onload="runTest();"></body>

</html>
